Added a +/- button to the cart page for easier add and removal of items. Also fixed search and store page titles.
